+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Add `ImageBuildOptsBuilder::unset_env`
 * Add `ManifestCreateOptsBuilder::amend`
 * Add `ManifestImageAddOptsBuilder::os_features`
+* Add `ManifestPushOptsBuilder::quiet`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync